Moves the get_or_create_patient function to the intrahospital_api.loader
