+++ intrahospital_api/loader.py
@@ -464,3 +464,21 @@ def _load_patient(patient, patient_load):
         patient_load.failed()
     else:
         patient_load.complete()
+
+
+def get_or_create_patient(mrn, episode_category, run_async=None):
+    patient = Patient.objects.filter(
+        demographics__hospital_number=mrn
+    ).first()
+    if not patient:
+        patient = Patient.objects.filter(
+        mergedmrn__mrn=mrn
+    ).first()
+
+    if patient:
+        patient.episode_set.get_or_create(
+            category_name=episode_category.display_name
+        )
+        return (patient, False)
+    patient = create_rfh_patient_from_hospital_number(mrn, episode_category, run_async=run_async)
+    return patient, True
