+++ intrahospital_api/test/test_loader.py
@@ -11,6 +11,7 @@ from intrahospital_api import loader
 from intrahospital_api.exceptions import BatchLoadError
 from intrahospital_api.constants import EXTERNAL_SYSTEM
 from plugins.labtests import models as lab_test_models
+from plugins.tb import episode_categories as tb_episode_categories
 
 
 @override_settings(API_USER="ohc")
@@ -850,3 +851,62 @@ class CreateRfhPatientFromHospitalNumberTestCase(OpalTestCase):
                 )
             ).exists()
         )
+
+
+class GetOrCreatePatientTestCase(OpalTestCase):
+    def setUp(self):
+        self.patient, _ = self.new_patient_and_episode_please()
+
+    def test_get_existing_patient(self):
+        self.patient.demographics_set.update(hospital_number="123")
+        self.patient.episode_set.update(
+            category_name=episode_categories.InfectionService.display_name
+        )
+        patient, created = loader.get_or_create_patient(
+            '123', episode_categories.InfectionService
+        )
+        self.assertEqual(self.patient, patient)
+        episode = self.patient.episode_set.get()
+        self.assertEqual(
+            episode.category_name,
+            episode_categories.InfectionService.display_name
+        )
+        self.assertFalse(created)
+
+    def test_create_new_episode_on_existing_patient(self):
+        self.patient.demographics_set.update(hospital_number="123")
+        patient, created = loader.get_or_create_patient(
+            '123', tb_episode_categories.TbEpisode
+        )
+        self.assertEqual(self.patient, patient)
+        self.assertTrue(self.patient.episode_set.filter(
+            category_name=tb_episode_categories.TbEpisode.display_name
+        ).exists())
+        self.assertFalse(created)
+
+    @mock.patch('intrahospital_api.loader.create_rfh_patient_from_hospital_number')
+    def test_get_merged_patient(self, create_rfh_patient_from_hospital_number):
+        self.patient.demographics_set.update(hospital_number="234")
+        self.patient.mergedmrn_set.create(mrn="123")
+        patient, created = loader.get_or_create_patient(
+            '123', episode_categories.InfectionService
+        )
+        self.assertEqual(
+            patient.id, self.patient.id
+        )
+        self.assertFalse(create_rfh_patient_from_hospital_number.called)
+        self.assertFalse(created)
+
+    @mock.patch('intrahospital_api.loader.create_rfh_patient_from_hospital_number')
+    def test_create_new_patient(self, create_rfh_patient_from_hospital_number):
+        create_rfh_patient_from_hospital_number.return_value = self.patient
+        patient, created = loader.get_or_create_patient(
+            '123', episode_categories.InfectionService
+        )
+        create_rfh_patient_from_hospital_number.assert_called_once_with(
+            '123',
+            episode_categories.InfectionService,
+            run_async=None
+        )
+        self.assertEqual(self.patient, patient)
+        self.assertTrue(created)
